Rework into core & rename lots of functions
