+++ source_files/QT_engine.pri
@@ -107,7 +107,7 @@ HEADERS += engine/player/expansion_effects.hpp
 HEADERS += engine/player/gear_stats.hpp
 HEADERS += engine/player/ground_aoe.hpp
 HEADERS += engine/player/instant_absorb.hpp
-HEADERS += engine/player/parse_player_effects.hpp
+HEADERS += engine/player/parse_auras.hpp
 HEADERS += engine/player/pet.hpp
 HEADERS += engine/player/pet_spawner.hpp
 HEADERS += engine/player/pet_spawner_impl.hpp
