+++ source_files/VS_engine.props
@@ -111,7 +111,7 @@ To change the list of source files run synchronize.py
 		<ClInclude Include="..\engine\player\gear_stats.hpp" />
 		<ClInclude Include="..\engine\player\ground_aoe.hpp" />
 		<ClInclude Include="..\engine\player\instant_absorb.hpp" />
-		<ClInclude Include="..\engine\player\parse_player_effects.hpp" />
+		<ClInclude Include="..\engine\player\parse_auras.hpp" />
 		<ClInclude Include="..\engine\player\pet.hpp" />
 		<ClInclude Include="..\engine\player\pet_spawner.hpp" />
 		<ClInclude Include="..\engine\player\pet_spawner_impl.hpp" />
