+++ source_files/cmake_engine.txt
@@ -105,7 +105,7 @@ player/expansion_effects.hpp
 player/gear_stats.hpp
 player/ground_aoe.hpp
 player/instant_absorb.hpp
-player/parse_player_effects.hpp
+player/parse_auras.hpp
 player/pet.hpp
 player/pet_spawner.hpp
 player/pet_spawner_impl.hpp
